\name{hett-internal}
\alias{dof.profile}
\title{
Internal profile likelihood function 
}
\description{
These are internal \code{hett} functions that are not intended to be
directly called by the user}
\usage{
dof.profile(dof, n, sqResid, orthoI, X, Z)
}
\author{Julian Taylor}
\examples{}
\keyword{internal}
